+++ microgen-codegen/src/test/java/org/openapitools/codegen/languages/TemplateType.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.v3.generators.java;
+package org.openapitools.codegen.languages;
 
 public enum TemplateType {
